\name{Boomerang-package}
\alias{Boomerang-package}
\alias{Boomerang}
\docType{package}
\title{
\packageTitle{Boomerang}
}
\description{
\packageDescription{Boomerang}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Boomerang}
\packageIndices{Boomerang}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{Boomerang}

%Maintainer: \packageMaintainer{Sean Devlin: devlins@mskcc.org}
}
\references{
%Devlin SM, Ostrovnaya I, and GonenM. Boomerang: A Method for Recursive Reclassication. Biometrics, accepted 
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
#~~ simple examples of the most important functions ~~
}
